Image-Classifier/app/requirements.txt
fastapi
uvicorn
tensorflow
pydantic
sqlalchemy
python-dotenv
requests
pillow